pyproject.toml
[tool.poetry]
name = "performance_tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
celery = "^5.2.7"
django-redis = "^5.2.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.2.0"
pytest-django = "^4.5.2"
model-bakery = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
